add audio input file (uses ffmpeg module like video) ; split big FFMpegTransformerModule.php file
